requirements.txt
asgiref==3.2.10
cffi==1.14.3
cryptography==3.2.1
Django==3.1.2
djangorestframework==3.12.2
gunicorn==20.0.4
pycparser==2.20
pytz==2020.1
six==1.15.0
sqlparse==0.4.1
whitenoise==5.2.0
